requirements.txt
aiogram==3.4.1
sqlalchemy==2.0.27
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
loguru==0.7.2
python-dotenv==1.0.1
